requirements.txt
Django==3.2.18
djangorestframework==3.14.0
psycopg2==2.9.5
djoser==2.1.0